+++ idea/src/org/jetbrains/kotlin/idea/quickfix/createFromUsage/createCallable/CreateGetSetFunctionActionFactory.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2010-2016 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jetbrains.kotlin.idea.quickfix.createFromUsage.createCallable
+
+import org.jetbrains.kotlin.KtNodeTypes
+import org.jetbrains.kotlin.diagnostics.Diagnostic
+import org.jetbrains.kotlin.diagnostics.Errors
+import org.jetbrains.kotlin.psi.KtArrayAccessExpression
+import org.jetbrains.kotlin.psi.KtBinaryExpression
+import org.jetbrains.kotlin.psi.KtContainerNode
+import org.jetbrains.kotlin.psi.KtPsiUtil
+
+abstract class CreateGetSetFunctionActionFactory(private val isGet: Boolean) : CreateCallableMemberFromUsageFactory<KtArrayAccessExpression>() {
+    override fun getElementOfInterest(diagnostic: Diagnostic): KtArrayAccessExpression? {
+        return when (diagnostic.factory) {
+            Errors.NO_GET_METHOD -> if (isGet) Errors.NO_GET_METHOD.cast(diagnostic).psiElement else null
+            Errors.NO_SET_METHOD -> if (!isGet) Errors.NO_SET_METHOD.cast(diagnostic).psiElement else null
+            in Errors.TYPE_MISMATCH_ERRORS -> {
+                val indicesNode = diagnostic.psiElement.parent as? KtContainerNode ?: return null
+                if (indicesNode.node.elementType != KtNodeTypes.INDICES) return null
+                val arrayAccess = indicesNode.parent as? KtArrayAccessExpression ?: return null
+                val parent = arrayAccess.parent
+                val isAssignmentLHS = KtPsiUtil.isAssignment(parent) && (parent as KtBinaryExpression).left == arrayAccess
+                if (isAssignmentLHS == isGet) return null
+
+                arrayAccess
+            }
+            else -> throw IllegalArgumentException("Unexpected diagnostic: ${diagnostic.factory.name}")
+        }
+    }
+}